Move tapErrors/errorTo to eigher package
